+++ src/private/cli/GetNovaCliInstalledVersion.ps1
@@ -1,0 +1,7 @@
+function Get-NovaCliInstalledVersion {
+    [CmdletBinding()]
+    param()
+
+    return $ExecutionContext.SessionState.Module.Version.ToString()
+}
+